+++ src/Facades/FaceBook.php
@@ -7,7 +7,8 @@ use Illuminate\Support\Facades\Facade;
 /**
  * Class FaceBook
  *
- * @method static mixed share(string $caption, string $url)
+ * @method static mixed shareUrl(string $caption, string $url)
+ * @method static mixed shareText(string $caption)
  * @method static mixed shareImage(string $caption, string $image_url)
  * @method static mixed shareVideo(string $caption, string $video_url)
  * @method static mixed getPageInsights(array $metrics = [], array $additionalParams = [])