+++ src/Facades/Telegram.php
@@ -6,7 +6,8 @@ use Illuminate\Support\Facades\Facade;
 /**
  * Class Telegram
  *
- * @method static mixed share(string $caption, string $url)
+ * @method static mixed shareUrl(string $caption, string $url)
+ * @method static mixed shareText(string $caption)
  * @method static mixed shareImage(string $caption, string $image_url)
  * @method static mixed shareDocument(string $caption, string $document_url)
  * @method static mixed shareVideo(string $caption, string $video_url)