+++ src/Services/InstagramService.php
@@ -359,7 +359,7 @@ class InstagramService extends SocialMediaService implements ShareInterface, Sha
      * @return array Response from the Instagram API.
      * @throws SocialMediaException
      */
-    public function share(string $caption, string $url): array
+    public function shareUrl(string $caption, string $url): array
     {
         $this->validateInput($caption, $url);